Teacher list Combobox holds generic placeholders for teachers, Teacher list Combobox is invisible on window launch. teacher list combobox becomes visible and useable when a course is selected.
removed all instances of "A Number" and replaced the functionality it held
with the name variable

Implemented/integrated teacher string across the application to be accepted
and properly displayed/transmitted.
modified all relevant inputs, outputs, variables, arguments etc... that involve
teacher or A-Number

error handling:
the mainWindow will no longer continue to execute the rest of the inserting
process unless the user directly clicks submit.

the mainWindow will no longer continue to excute the rest of the tutor completing
process unless the tutor specifically clicks submit.

attempting to submit a tutor completion without a unique description,
then exiting that window is no longer a bug to bypasse the "continuing" check.

You can no longer click on a completed user and fill out/submit a tutor form
for them again.

All text entered into the database is now standardized to all upper case, so names
of teachers students and courses will not be mismatched from other entries.